projet-multi-omiques/requirements.txt
# Dépendances principales
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0

# Pour les tests
pytest>=6.2.0
pytest-cov>=2.12.0

# Pour les notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Pour le traitement des données
scipy>=1.7.0

# Pour le matching fuzzy (optionnel)
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0